+++ crates/ptool/src/lib.rs
@@ -9,6 +9,7 @@ mod re;
 mod runner;
 mod script_args;
 mod semver;
+mod ssh;
 mod template;
 mod text;
 mod toml;