+++ crates/ptool/src/lua_api.rs
@@ -40,6 +40,7 @@ fn create_ptool_module(
     let shell_module = create_ptool_shell_module(lua, Rc::clone(&world))?;
     let http_module = create_ptool_http_module(lua, Rc::clone(&world))?;
     let db_module = create_ptool_db_module(lua, Rc::clone(&world))?;
+    let ssh_module = create_ptool_ssh_module(lua, Rc::clone(&world))?;
     let fs_module = create_ptool_fs_module(lua, Rc::clone(&world))?;
     let path_module = create_ptool_path_module(lua, Rc::clone(&world))?;
     let re_module = create_ptool_re_module(lua, Rc::clone(&world))?;
@@ -55,6 +56,7 @@ fn create_ptool_module(
     module.set("sh", shell_module)?;
     module.set("http", http_module)?;
     module.set("db", db_module)?;
+    module.set("ssh", ssh_module)?;
     module.set("fs", fs_module)?;
     module.set("path", path_module)?;
     module.set("re", re_module)?;
@@ -117,6 +119,14 @@ fn create_ptool_db_module(lua: &Lua, world: Rc<RefCell<crate::LuaWorld>>) -> mlu
         lua.create_function(move |_, value: Value| world.borrow().db_connect(value))?;
     db_module.set("connect", connect_fn)?;
     Ok(db_module)
+}
+
+fn create_ptool_ssh_module(lua: &Lua, world: Rc<RefCell<crate::LuaWorld>>) -> mlua::Result<Table> {
+    let ssh_module = lua.create_table()?;
+    let connect_fn =
+        lua.create_function(move |_, value: Value| world.borrow().ssh_connect(value))?;
+    ssh_module.set("connect", connect_fn)?;
+    Ok(ssh_module)
 }
 
 fn create_ptool_fs_module(lua: &Lua, world: Rc<RefCell<crate::LuaWorld>>) -> mlua::Result<Table> {